build: move build/utest folders
