+++ .eslintrc
@@ -1,5 +1,5 @@
 {
-  "extends": "plugin:@essex/release",
+  "extends": "@essex/eslint-config/release",
   "rules": {
     "@typescript-eslint/no-explicit-any": "off" 
   }
