+++ packages/scripts/src/commands/watch/types.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { BundleMode } from '../build/types'
+import { BundleMode } from '../_build/types'
 
 export interface WatchCommandOptions {
 	verbose?: boolean
