+++ packages/steps/eslint/config/.eslintrc-experiment
@@ -1,3 +1,3 @@
 {
-    "extends": ["plugin:@essex/experiment"]
+    "extends": ["@essex/eslint-config/experiment"]
 }
