+++ packages/steps/eslint/config/.eslintrc-release
@@ -1,3 +1,3 @@
 {
-    "extends": ["plugin:@essex/release"]
+    "extends": ["@essex/eslint-config/release"]
 }
